requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector,speechify,google]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn